+++ src/server/routes/services/{serviceId}/diagnostics/dev/dashboards/promote/page.njk
@@ -1,0 +1,30 @@
+{% extends "layouts/page.njk" %}
+
+{% block content %}
+  <div class="govuk-main-wrapper app-main-wrapper">
+
+    {{ appPageHeading({
+      caption: "Promote",
+      text: "Dashboards",
+      intro: 'Promote your playground dashboards'
+    }) }}
+
+    <div class="govuk-grid-row">
+      <div class="govuk-grid-column-one-third-from-desktop-wide">
+
+      {{ govukTable({
+        classes: "app-table govuk-!-margin-bottom-4",
+        head: [
+          { text: "Type" },
+          { text: "Dashboard" },
+          { text: "Version" }
+        ],
+        rows: []
+      }) }}
+
+
+      </div>
+    </div>
+
+  </div>
+{% endblock %}
