+++ src/server/routes/services/{serviceId}/diagnostics/utils/transformResources.js
@@ -1,0 +1,11 @@
+export default function transformResources(resources = {}) {
+  return Object.fromEntries(
+    Object.entries(resources).map(([key, value]) => {
+      if (key === 'metrics') {
+        return [key, { ...Object.groupBy(value, (item) => item.type) }]
+      }
+
+      return [key, value]
+    })
+  )
+}
